Introduce a set of master keys for the store.
Rather than directly encrypting each item using the password, this
introduces a set of master keys that encrypt each item, and which
themselves are encrypted with the password. They are generated at random
when the store is first used. This means that getting an item requires
an additional request to get the keys, but has several benefits.

It prevents multiple items being written with different passwords. The
current design allows multiple items to be written with different
passwords, since writing one does not require reading the other.
Requiring the second write to decrypt the master keys prevents this.

It also means changing the store password only requires changing one
file, rather than re-encrypting every item in the store.